+++ packages/nx-plugin/src/generators/app/lib/add-eslint.ts
@@ -13,12 +13,12 @@ export async function addEslint(
     eslintFilePatterns: [`${options.projectRoot}/**/*.{ts,html}`],
     skipFormat: true,
   };
-  if (majorNxVersion === 16) {
+  if (majorNxVersion >= 16) {
     await (
       await import(
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore
-        '@nx/linter'
+        '@nx/eslint'
       )
     ).lintProjectGenerator(tree, linterOptions);
   } else {