+++ packages/vite-plugin-nitro/src/lib/vite-plugin-nitro.ts
@@ -1,5 +1,5 @@
 import type { NitroConfig } from 'nitropack';
-import { toNodeListener } from 'h3';
+import { App, toNodeListener } from 'h3';
 import type { Plugin, UserConfig } from 'vite';
 import { normalizePath, ViteDevServer } from 'vite';
 import * as path from 'path';
@@ -138,11 +138,17 @@ export function nitro(options?: Options, nitroOptions?: NitroConfig): Plugin[] {
                 },
               ],
               externals: {
-                external: ['rxjs', 'node-fetch-native/dist/polyfill', 'destr'],
+                ...nitroOptions?.externals,
+                external: [
+                  'rxjs',
+                  'node-fetch-native/dist/polyfill',
+                  ...(nitroOptions?.externals?.external || []),
+                ],
               },
               moduleSideEffects: [
-                'zone.js/plugins/zone-node',
+                'zone.js/node',
                 'zone.js/fesm2015/zone-node',
+                ...(nitroOptions?.moduleSideEffects || []),
               ],
               renderer: normalizePath(`${__dirname}/runtime/renderer`),
               handlers: [
@@ -162,8 +168,9 @@ export function nitro(options?: Options, nitroOptions?: NitroConfig): Plugin[] {
       },
       async configureServer(viteServer: ViteDevServer) {
         if (isServe && !isTest) {
-          const { createNitro, createDevServer, build, prepare } =
-            await loadEsmModule<typeof import('nitropack')>('nitropack');
+          const { createNitro, createDevServer, build } = await loadEsmModule<
+            typeof import('nitropack')
+          >('nitropack');
 
           const nitro = await createNitro({
             dev: true,
@@ -171,7 +178,17 @@ export function nitro(options?: Options, nitroOptions?: NitroConfig): Plugin[] {
           });
           const server = createDevServer(nitro);
           await build(nitro);
-          viteServer.middlewares.use(apiPrefix, toNodeListener(server.app));
+          viteServer.middlewares.use(
+            apiPrefix,
+            toNodeListener(server.app as unknown as App)
+          );
+
+          viteServer.httpServer?.once('listening', () => {
+            process.env['ANALOG_HOST'] = !viteServer.config.server.host
+              ? 'localhost'
+              : (viteServer.config.server.host as string);
+            process.env['ANALOG_PORT'] = `${viteServer.config.server.port}`;
+          });
 
           console.log(
             `\n\nThe server endpoints are accessible under the "${apiPrefix}" path.`